+++ python/taichi/aot/_export.py
@@ -1,0 +1,11 @@
+import taichi as ti
+
+_aot_kernels = []
+
+
+def export(f):
+    assert hasattr(f,
+                   "_is_wrapped_kernel"), "Only Taichi kernels can be exported"
+    out = f
+    _aot_kernels.append(out)
+    return out
